fix(gate-api/pool-api): fixed deployment workflow, ParameterObjectArgumentResolver class moved to common module
